Re-added unused `set;` accessor for serialization
