+++ src/NCalc/Domain/BinaryExpression.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.CodeAnalysis;
 using NCalc.Visitors;
 
 namespace NCalc.Domain;
@@ -7,11 +8,11 @@ public sealed class BinaryExpression(
         LogicalExpression leftExpression,
         LogicalExpression rightExpression) : LogicalExpression
 {
-    public LogicalExpression LeftExpression { get;  } = leftExpression;
+    public LogicalExpression LeftExpression { get; set; } = leftExpression;
 
-    public LogicalExpression RightExpression { get;  } = rightExpression;
+    public LogicalExpression RightExpression { get; set;  } = rightExpression;
 
-    public BinaryExpressionType Type { get; } = type;
+    public BinaryExpressionType Type { get; set;  } = type;
 
     public override void Accept(LogicalExpressionVisitor visitor)
     {
