+++ src/NCalc/Domain/Function.cs
@@ -4,9 +4,9 @@ namespace NCalc.Domain;
 
 public sealed class Function(Identifier identifier, LogicalExpression[] expressions) : LogicalExpression
 {
-    public Identifier Identifier { get;  } = identifier;
+    public Identifier Identifier { get; set; } = identifier;
 
-    public LogicalExpression[] Expressions { get; } = expressions;
+    public LogicalExpression[] Expressions { get; set; } = expressions;
 
     public override void Accept(LogicalExpressionVisitor visitor)
     {
