+++ src/NCalc/Domain/UnaryExpression.cs
@@ -1,14 +1,12 @@
-using System;
 using NCalc.Visitors;
 
 namespace NCalc.Domain;
 
-[Serializable]
 public sealed class UnaryExpression(UnaryExpressionType type, LogicalExpression expression) : LogicalExpression
 {
-    public LogicalExpression Expression { get; } = expression;
+    public LogicalExpression Expression { get; set;  } = expression;
 
-    public UnaryExpressionType Type { get; } = type;
+    public UnaryExpressionType Type { get; set;  } = type;
 
     public override void Accept(LogicalExpressionVisitor visitor)
     {
